feat: experimental python packaging and interface
