+++ tgi/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "tgi"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "tgi"
+crate-type = ["cdylib"]
+
+[dependencies]
+pyo3 = { version = "0.20.0", features = ["extension-module"] }
+pyo3-asyncio = { version = "0.20.0", features = ["tokio-runtime"] }
+tokio = "1.4"
+text-generation-router = { path = "../router" }
+text-generation-launcher = { path = "../launcher" }
