+++ tgi/Makefile
@@ -1,0 +1,6 @@
+
+build:
+	maturin build
+
+install: build
+	pip install -e .
